requirements.txt
asgiref==3.9.1
Django==5.2.6
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
